requirements_dev.txt
# core
colorama==0.4.4
coverage==5.5

# dev
black==21.6b0
commitizen==2.17.11
flake8==3.9.2
invoke==1.5.0
isort==5.9.1
mypy==0.910
pre-commit==2.13.0

# test
codecov==2.1.11
faker==8.9.1
pytest-cov==2.12.1
pytest-mock==3.6.1
pytest-sugar==0.9.4
tox==3.23.1

# dist
build==0.5.1
twine==3.4.1
